requirements.dev.txt
pynose>=1.5.4
mock>=5.2.0
coverage>=7.7.0
flake8>=7.1.2
tox>=4.24.2
